modified testAnt method to accept boolean array, made Operations class for users to fill their methods
