+++ src/main/java/com/Ant.java
@@ -1,3 +1,5 @@
+package com;
+
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
