+++ src/test/java/GameTest.java
@@ -1,3 +1,5 @@
+import com.Ant;
+import com.GameField;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
